requirements.txt
APScheduler==3.11.2
PyJWT==2.10.1
PyYAML==6.0.3
aioredis==2.0.1
argon2-cffi==23.1.0
bcrypt==5.0.0
fastapi==0.128.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.4
python-dotenv==1.0.0
python-multipart==0.0.6
redis==7.1.0
scikit-learn==1.8.0
starlette==0.50.0
uvicorn==0.40.0
webauthn==2.2.0
xgboost==3.1.3

# Additional dependencies
groq==0.18.0
requests==2.32.0
aiohttp==3.11.0
matplotlib==3.10.0
seaborn==0.13.2
